+++ Projects/Feature/Sent/Sources/SpecificEnvelope/SpecificEnvelopeHistoryDetail/EnvelopeDetailProperty.swift
@@ -10,9 +10,9 @@ import Foundation
 
 struct EnvelopeDetailProperty: Equatable, Identifiable {
   /// 봉투의 아이디 입니다.
-  let id: Int
+  let id: Int64
   /// 현재 봉투의 가격을 나타냅니다.
-  let price: Int
+  let price: Int64
   /// 현재 봉투의 경조사 이름을 나타냅니다.
   let eventName: String
   /// 봉투를 받은 사람의 이름을 나타냅니다.
@@ -87,7 +87,7 @@ struct EnvelopeDetailProperty: Equatable, Identifiable {
   ///   - relation: 현재 봉투를 주고받은 사람과의 관계를 나타냅니다.
   ///   - date: 현재 봉투를 주고받은 날짜를 나타냅니다.
   ///   - isVisited: 현재 봉투의 대상이되는 경조사에 참석 여부를 나타냅니다.
-  init(id: Int, price: Int, eventName: String, name: String, relation: String, date: Date, isVisited: Bool?, gift: String? = nil, contacts: String? = nil, memo: String? = nil) {
+  init(id: Int64, price: Int64, eventName: String, name: String, relation: String, date: Date, isVisited: Bool?, gift: String? = nil, contacts: String? = nil, memo: String? = nil) {
     self.id = id
     self.price = price
     self.eventName = eventName
